-Los métodos "hayColision" y "getRebote" del tipo "Raqueta" reciben ahora un puntero del tipo "Pelota".
-Se ha cambiado el nombre del atributo tamanio por alto, así como los métodos getTamanio por getAlto y el orden en el que recibe los atributos el constructor.

-Modificado el método muevete de "PelotaNormal" (sin terminar), ahora se trata de un método recursivo que calculará todos los posibles choques en una iteración. Requiere métodos de "Raqueta" y "Espacio" que devuelvan el tiempo(porción de iteración) que tardará en colisionar(método por hacer).


git-svn-id: http://pong-opengl.googlecode.com/svn/trunk@26 841c6515-18fd-ad39-5d5d-772e54db8338

+++ tipos/src/RaquetaNormal.cpp
@@ -49,8 +49,8 @@ void RaquetaNormal::incrPuntos(){ puntuacion++; }
 TipoColision RaquetaNormal::hayColision(Vector posicion, int radio) const {
 	
 	TipoColision tColision;
-	int x = posicion.x;
-	int y = posicion.y;
+	double x = posicion.x;
+	double y = posicion.y;
 
 	if(
 		(
